# specs/00008-hybrid-retrieval-and-reranking/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Hybrid Retrieval and Reranking API
  version: 1.0.0
  summary: Ranked passages from a fused lexical and dense retrieval path, reranked by a local cross-encoder.
  description: |
    Request-serving boundary for E008. Three operations: one retrieval query, one readiness probe, one
    diagnostics report. Two consumers, neither of them a browser — E011 (grounded chat) calls
    `GET /retrieval/search` for passages to answer from, and E014 (evaluation) calls the same operation
    once per arm to build an ablation table, reading the ranking parameters this surface returns with
    every result.

    **No write path.** E008 reads the chunk store and writes nothing to it — that exclusion is stated in
    the specification's Scope and includes `chunk.part_numbers`. This contract therefore defines no POST,
    PUT, PATCH or DELETE anywhere, which makes the read-only property a fact about the route table
    rather than a discipline someone maintains.

    **`GET` rather than `POST` for the query.** The request is a pure read with a flat scalar parameter
    set — a query string, an arm selection, an optional project filter, a result bound. The one existing
    endpoint on this boundary (`GET /api/v1/worklist`) established the convention and this follows it.
    The reversal trigger is a request that stops being flat: a structured filter object, or a query
    routinely past the 1000-character cap declared on `query`. Neither exists today.

...
    `POST`.

...
servers:
...
paths:
  /retrieval/search:
    get:
      tags: [retrieval]
      operationId: searchRetrieval
      summary: Ranked passages for a query, with their provenance, arms, and the ranking parameters in force.
      description: |
        The whole retrieval path in one response: the deterministic route, the fused lexical and dense
        arms, and reranking, with every result carrying the document identity, document type, project
        and page number **read from the chunk** (FR-008).

        **Order of execution**, and it is load-bearing: the deterministic route resolves part-number
        shaped tokens first (FR-010) and falls through on a miss (FR-011); the lexical and dense arms are
        fetched and fused inside one database statement (FR-002); the fused candidate set is reranked in
        full (FR-018). Nothing here re-orders on the application side using a score this response does
        not carry.

        **Determinism.** For an identical query against an unchanged corpus and an unchanged
        configuration the ordering is identical (FR-020) and `ordering_digest` is equal. On the
        approximate path that guarantee is scoped to a fixed index: graph construction is randomized, so
        a rebuild can legitimately change which approximate candidates are returned. This contract states
        the property; the plan owns the scoping (see § Requirements this contract cannot express).
      parameters:
        - $ref: '#/components/parameters/QueryText'
        - $ref: '#/components/parameters/Arm'
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: |
            Ranked passages, possibly none. An empty result is a fully successful answer: `results` is
            `[]`, `result_count` is `0`, and no low-scoring filler is added to reach a count (FR-009).
            A degraded response — fusion-only because the reranker is unavailable — is also a `200`,
            distinguished by `mode`, never by a status code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalResponse'
              examples:
                reranked:
                  $ref: '#/components/examples/RerankedSearch'
                deterministicRoute:
                  $ref: '#/components/examples/DeterministicRouteSearch'
                degraded:
                  $ref: '#/components/examples/DegradedSearch'
                empty:
                  $ref: '#/components/examples/EmptySearch'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatastoreUnavailable'
...
components:
  parameters:
    QueryText:
      name: query
      in: query
      required: true
      description: |
        The coordinator's question, or a bare identifier, verbatim.

        **No transformation of any kind is applied.** The specification excludes query rewriting,
        expansion and hypothetical-document generation, with the measured evidence and the reversal
        trigger recorded in Scope § Excluded. This string is what the lexical arm matches, what the
        encoder embeds, and what the deterministic route scans for part-number-shaped tokens. A caller
        must not pre-transform it either: a rewrite substitutes vocabulary in 95% of cases and damages
        the arm carrying the part numbers hardest.

        The 1000-character cap is set by this contract, not derived from a requirement. It keeps the
        request well inside URL length limits and is the boundary at which the `GET`-versus-`POST`
        decision above would be revisited. A query at or under the cap is still truncated by the encoder
        at its own word-piece limit; see § Requirements this contract cannot express for why that
        truncation is not reported here.

        Empty or whitespace-only is a `422`, not an empty result: a query with nothing in it and a query
        that matched nothing are different statements and FR-009 governs only the second.
      schema:
        type: string
        minLength: 1
        maxLength: 1000
      example: what does the spec require for galvanizing repair
    Arm:
      name: arm
      in: query
      required: false
      description: |
        FR-025. Which retrieval arm runs. Each returns results without requiring the others, which is
        what makes E014's ablation table constructible: without a request-level selection there is no way
        to obtain a lexical-only or a fusion-only ranking from a running service, and no other epic can
        construct an arm this one did not make runnable.

        Five values, enumerated machine-readably at the root `x-arms` extension. The sixth arm SC-012
        names — approximate vector search — is **not** here: it is the FR-026 configuration flag, read at
        process start and reported in `ranking_parameters.vector_search`. See `info` § The
        exact-versus-approximate flag for why admitting it as a request field would be the drift ADR-0005
        conditions its acceptance on preventing.

        `fused_reranked` is the default and the served path, so E011 selects nothing.
      schema:
        $ref: '#/components/schemas/Arm'
    ProjectId:
      name: project_id
      in: query
      required: false
      description: |
        Scopes retrieval to one project. Both arms and the deterministic route are filtered alike; the
        chunk store carries `ix_chunk__project` for exactly this.

        FR-028 binds here: a filtered vector search MUST return the requested candidate count wherever
        that many matching rows exist, rather than the smaller set the graph index returns when it
        applies the filter after selecting candidates. A filtered query that yields fewer than
        `ranking_parameters.fetch_depth` dense candidates while more matching rows exist is a defect
        under SC-014, not a property of the corpus — which is why `fused_candidate_count` is returned.

        A well-formed identifier matching no chunk returns `200` with an empty result, never `404`.
        There is no project registry to consult; `project_id` is a format-checked column, so the system
        cannot distinguish "no such project" from "a project with no chunks", and reporting a
        distinction the data cannot support is the defect FR-009 refuses in the other direction.
      schema:
        type: string
        pattern: '^PRJ-[0-9]{3}$'
      example: PRJ-002
    Limit:
      name: limit
      in: query
      required: false
      description: |
        Bounds the number of **ranked-relevance** results returned. Deterministic identifier matches are
        additional to this bound and are never displaced by it (`info` § Recorded interface decisions,
        item 2), because bounding the whole list would let a route hit at position 1 push a fused
        candidate out of the response — FR-012's "the route MUST NOT remove" violated by an interaction
        rather than by a rule.

        **It cannot change the ranking.** Fusion fetches 50 candidates per arm (FR-003) and reranking
        scores exactly the fused candidate set (FR-018) whatever this value is; truncation happens after
        both. `reranking.candidates_scored` therefore does not move with `limit`, and that is assertable
        from the response.

        **It is not padding.** FR-009 forbids padding a short result set to reach a target count, and
        this is a ceiling, never a floor: `result_count` is what was found, up to this bound.

        E014 must record it, because recall@k read from a response truncated at a smaller `k` is a
        different figure. It is echoed in `request.limit` for that reason.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 50
      example: 8
...
  schemas:
...
    RetrievalResponse:
      type: object
      description: |
        **Counting invariants**, all assertable from one response:

          - `result_count == length(results)`;
          - `result_count == min(limit, ranked_relevance_count) + deterministic_route.added_count`,
            where `ranked_relevance_count` is the number of fused candidates the route did not promote —
            `limit` bounds the ranked-relevance portion only (`info` § Recorded interface decisions);
          - `fused_candidate_count <= ranking_parameters.fetch_depth * 2` and, where reranking ran,
            `reranking.candidates_scored == fused_candidate_count` (FR-018);
          - every result with a non-null `fused_rank` is a candidate hybrid retrieval produced, so
            FR-012's "the route MUST NOT remove" is checkable without a route-disable switch;
          - **`result_count` is never raised to reach a target.** FR-009: an empty result set is reported
            as empty and a short one is not padded. `results: []` with `result_count: 0` is a complete,
            successful answer.
      additionalProperties: false
      required:
        - generated_at
        - request
        - mode
        - deterministic_route
        - ranking_parameters
        - reranking
        - results
        - result_count
        - fused_candidate_count
        - ordering_digest
      properties:
        generated_at:
          type: string
          format: date-time
          description: |
            When this response was produced. The single member two responses to an identical query
            against an unchanged corpus and configuration may differ in — it describes the act of
            answering rather than what was answered, and takes no part in `ordering_digest` for that
            reason.
        request:
          $ref: '#/components/schemas/RequestEcho'
        mode:
          $ref: '#/components/schemas/Mode'
        deterministic_route:
          $ref: '#/components/schemas/DeterministicRoute'
        ranking_parameters:
          $ref: '#/components/schemas/RankingParameters'
        reranking:
          oneOf:
            - $ref: '#/components/schemas/RerankingReport'
            - type: 'null'
          description: Null exactly when `mode.reranked` is false.
        results:
          type: array
          description: |
            The final ordering. Deterministic identifier matches first, in the tie-break key's order;
            then the ranked-relevance results in reranked order, or in fused order where reranking did
            not run. Array position is authoritative and `rank` restates it.

            **`maxItems` is a contract-originated bound and only half of it is derived** *(disclosed
            2026-07-29)*. The ranked-relevance portion is bounded at 50 by `limit`. FR-037 bounds the
            *fused candidate set* at 100 — two arms of 50 fuse to at most 100 distinct candidates — but
            that is not this array: `limit` cuts the ranked portion before it is returned, and the
            deterministic route's additions are counted outside `limit` by design (`info` § Recorded
            interface decisions, item 2). **No requirement bounds the route's match count**, so the
            residual 50 is a number this contract chose. Either the route's contribution acquires a
            stated ceiling with a stated behaviour when it is reached, or this bound is dropped and the
            array is declared unbounded above `limit`. That is a decision nobody has made; it is
            recorded in `plan.md` § Contract Decisions Confirmed at Plan rather than settled here.
          maxItems: 100
          items:
            $ref: '#/components/schemas/RetrievalResult'
        result_count:
          type: integer
          minimum: 0
          description: |
            `length(results)`, restated so a consumer storing a truncated record still holds the count
            the response asserted. Zero is an honest, complete answer (FR-009).
        fused_candidate_count:
          type: integer
          minimum: 0
          description: |
            Candidates the fusion statement produced, before promotion by the deterministic route and
            before `limit`. Returned because it is the figure SC-014 is read against: a filtered query
            yielding fewer candidates than `fetch_depth` while more matching rows exist is recall lost to
            a retrieval setting rather than to the ranking design, and without this number that loss is
            invisible.
        ordering_digest:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
          description: |
            Digest over the ordered `chunk_id` sequence of `results`. FR-020 requires identical ordering
            for identical queries against an unchanged corpus and configuration, and SC-012 requires each
            arm to return identical results across two runs; this gives both a single assertable value
            instead of an element-wise comparison.

            **Domain.** Defined on every response, including one whose `results` is empty, where it is
            the digest of the empty sequence. Two responses that return nothing therefore carry the same
            digest however much else differs between them — equality of this value alone MUST NOT be read
            as sameness of response. It is not a cache validator and this surface defines none.

            On the approximate path the guarantee it witnesses holds against a fixed index. Graph
            construction is randomized, so a rebuild can legitimately move this value; the plan owns
            scoping FR-020's claim to the exact path and giving the approximate path an overlap and
            recall-delta assertion instead.
...
  examples:
...
    EmptySearch:
      summary: Nothing matched — reported as empty, never padded (FR-009)
      value:
        generated_at: '2026-07-29T10:33:08Z'
        request:
          query: seismic base isolation bearings for the turbine hall
          arm: fused_reranked
          project_id: PRJ-003
          limit: 8
        mode:
          arm_requested: fused_reranked
          arm_served: fused_reranked
          reranked: false
          degraded: false
          unreranked_reason: no_candidates_to_score
          statement: 'No candidate was retrieved, so there was nothing for the reranker to score. The empty result is the answer, not a shortfall to be filled.'
        deterministic_route:
          fired: false
          skipped_reason: null
          pattern: '\b[A-Z]{2,6}-[0-9]{3,8}\b'
          matched_tokens: []
          match_count: 0
          added_count: 0
          fell_through: false
        ranking_parameters:
          fusion_constant: 60
          tie_break_key: chunk_id_ascending
          missing_arm_convention: zero_contribution_from_the_absent_arm
          fetch_depth: 50
          vector_search:
            mode: approximate
            index:
              kind: hnsw
              ef_search: 100
              iterative_scan: strict_order
          lexical:
            ranking_basis: postgresql_tsvector_ranking
            uses_corpus_wide_term_statistics: false
            weights: [0.1, 0.2, 0.4, 1.0]
          encoder:
            model_id: sentence-transformers/all-MiniLM-L6-v2
            model_revision: '8b3219a92973c328a8e22fadcfa821b5dc75636a'
          reranker: null
        reranking: null
        results: []
        result_count: 0
        fused_candidate_count: 0
        ordering_digest: 'sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855'
...